Move Twitter to service, move loading to about page, add multiple widgets
