+++ components/TwitterTestimonials.vue
@@ -24,82 +24,19 @@
   </v-card>
 </template>
 <script>
-import testimonialsData from "../data/testimonials.json";
-import tweetFixtures from "../fixtures/twitterTweetsApiFixture.json";
-import logger from "../services/logger";
-
-// Axios requires proxies for origin changes. /twitter/ will get overwritten with https://api.twitter.com.
-const API_URL =
-  "/twitter/2/tweets?user.fields=profile_image_url,username&expansions=author_id&ids=";
-const DEV_MODE = "DEV_MODE";
-
-const token = process.env.TWITTER_API_BEARER_TOKEN || DEV_MODE;
-const options = {
-  headers: { Authorization: `Bearer ${token}` },
-  mode: "no-cors",
-};
-
 export default {
+  props: {
+    tweets: {
+      type: Array,
+      required: true
+    }
+  },
   data() {
     return {
-      tweets: [],
       activeTweetIndex: 0,
       isLoaded: false,
     };
-  },
-  methods: {
-    createUserMap: function (apiUserData) {
-      /* Maps users by id. */
-      const userMap = {};
-      for (const user of apiUserData) {
-        userMap[user.id] = user;
-      }
-      return userMap;
-    },
-    annotateTweetsWithUsers: function (apiTweetData, userMap) {
-      /* The twitter API separates user information and tweet content. This method recombines, overwriting fields as necessary. */
-      let annotated = [];
-      for (const tweet of apiTweetData) {
-        const annotatedTweet = Object.assign(tweet, {
-          name: userMap[tweet.author_id].name,
-          username: userMap[tweet.author_id].username,
-          profile_image_url: userMap[tweet.author_id].profile_image_url,
-          profile_image_url_original: userMap[
-            tweet.author_id
-          ].profile_image_url.replace(/_normal\./, "."),
-        });
-        annotated.push(annotatedTweet);
-      }
-      return annotated;
-    },
-  },
-  async fetch() {
-    /* Automatically called on component instantiation. */
-
-    let json = "";
-    if (token === DEV_MODE) {
-      logger.info(
-        "Twitter instantiated using dev mode; sample tweets will be loaded from fixtures."
-      );
-      json = tweetFixtures;
-    } else {
-      try {
-        const res = await this.$axios.get(
-          `${API_URL}${testimonialsData.tweets}`,
-          options
-        );
-        json = res.data;
-      } catch (e) {
-        logger.error("Could not retrieve tweets from Twitter API.", e.message);
-      }
-    }
-
-    const userMap = this.createUserMap(json.includes.users);
-    const annotatedTweets = this.annotateTweetsWithUsers(json.data, userMap);
-
-    this.tweets = annotatedTweets;
-    this.isLoaded = true;
-  },
+  }
 };
 </script>
 <style>
